refactor e fix: alterando nome de arquivo e função para português / corrigindo mensagem de erro
